+++ reportcard-model/src/main/java/com/ericdriggs/reportcard/model/ResultParserUtil.java
@@ -1,22 +1,22 @@
-package com.ericdriggs.reportcard.model.converter.surefire;
+package com.ericdriggs.reportcard.model;
 
-import com.ericdriggs.reportcard.model.TestResult;
+import com.ericdriggs.reportcard.model.converter.surefire.SurefireConvertersUtil;
 import com.ericdriggs.reportcard.xml.surefire.SurefireParserUtil;
 import com.ericdriggs.reportcard.xml.surefire.Testsuite;
 
 import java.util.List;
 
-public enum SurefireUtil {
+public enum ResultParserUtil {
 
     ;//static methods only
     private static final String XML_EXTENSION_REGEX = ".*[.]xml";
 
-    public static TestResult fromPath(String absolutePath) {
-        return fromAbsolutePath(absolutePath, XML_EXTENSION_REGEX);
+    public static TestResult fromSurefirePath(String absolutePath) {
+        return fromSurefirePathAndRegex(absolutePath, XML_EXTENSION_REGEX);
     }
 
 
-    public static TestResult fromAbsolutePath(String absolutePath, String fileNameRegex) {
+    public static TestResult fromSurefirePathAndRegex(String absolutePath, String fileNameRegex) {
         List<Testsuite> testsuites = SurefireParserUtil.parseTestSuitesFromPathAndRegex(absolutePath, fileNameRegex);
         return SurefireConvertersUtil.doFromSurefireToModelTestResult(testsuites);
     }
